Reorg package and slcaa name
