+++ eagle-core/eagle-app/eagle-app-base/src/main/java/org/apache/eagle/app/XmlSchedulingApplication.java
@@ -36,9 +36,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.text.ParseException;
 
-/**
- * TODO: Support load scheduling from XML instead of in-line code.
- */
 public class XmlSchedulingApplication extends SchedulingApplication {
     private static final Logger LOGGER = LoggerFactory.getLogger(XmlSchedulingPlan.class);
