+++ eagle-core/eagle-app/eagle-app-base/src/main/java/org/apache/eagle/app/job/plugins/XMLStreamSchedulingDataProcessor.java
@@ -52,7 +52,7 @@ public class XMLStreamSchedulingDataProcessor extends XMLSchedulingDataProcessor
     }
 
     public XMLStreamSchedulingDataProcessor() throws ParserConfigurationException {
-        super(new SimpleClassLoadHelper());
+        super(new ThreadContextClassLoadHelper());
     }
 
     public void processStreamAndUnscheduleJobs(InputStream stream, String systemId, Scheduler scheduler)
