+++ eagle-hadoop-health/src/main/java/org/apache/eagle/health/job/PropertiesUtils.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.health;
+package org.apache.eagle.health.job;
 
 import org.apache.hadoop.conf.Configuration;
 import org.slf4j.Logger;
