+++ eagle-hadoop-health/src/main/java/org/apache/eagle/health/job/URLPingCheckJob.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.health.detector;
+package org.apache.eagle.health.job;
 
 public class URLPingCheckJob {
