+++ eagle-hadoop-health/src/test/resources/application.conf
@@ -14,7 +14,7 @@
 # limitations under the License.
 
 {
-  appId = "HadoopHealthCheckAppTest"
+  appId = "HBASE_HEALTH_CHECK_APP_TESTSITE"
   siteId = "testsite"
 
   hbaseHealth {
